+++ objc2-encode/CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 * **BREAKING**: Sealed the `EncodeArguments` trait.
 
+### Removed
+* **BREAKING**: Removed `PartialEq` impl between `str` and `Encoding` since it
+  was incorrect (it violated the trait requirements).
+
 
 ## 2.0.0-beta.2 - 2022-01-03